feat: Replace UniversalSigner with UniversalAddress
