+++ primitives/runtime/src/lib.rs
@@ -26,21 +26,23 @@ mod universaladdress;
 pub use accountname::AccountName;
 pub use universaladdress::{UniversalAddress, UniversalAddressKind};
 
+#[cfg(feature = "serde")]
+pub use serde::{Deserialize, Serialize};
+
+use derive_more::{From, TryInto};
 use np_crypto::{p256, webauthn};
 use parity_scale_codec::{Decode, Encode};
 use scale_info::TypeInfo;
-use sp_core::{ecdsa, ed25519, sr25519};
+use sp_core::{crypto::AccountId32, ecdsa, ed25519, sr25519};
 use sp_runtime::{
 	traits::{IdentifyAccount, Lazy, Verify},
 	RuntimeDebug,
 };
-
-#[cfg(feature = "std")]
-use serde::{Deserialize, Serialize};
+use sp_std::prelude::*;
 
 /// Signature verify that can work with any known signature types.
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
 #[derive(Eq, PartialEq, Clone, Encode, Decode, RuntimeDebug, TypeInfo)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub enum UniversalSignature {
 	/// A Ed25519 signature.
 	Ed25519(ed25519::Signature),
@@ -55,87 +57,44 @@ pub enum UniversalSignature {
 }
 
 impl Verify for UniversalSignature {
-	type Signer = UniversalSigner;
+	type Signer = UniversalAddress;
 
-	fn verify<L: Lazy<[u8]>>(&self, mut msg: L, signer: &UniversalAddress) -> bool {
+	fn verify<L: Lazy<[u8]>>(&self, mut msg: L, signer: &AccountId32) -> bool {
 		match (self, signer) {
-			(Self::Ed25519(ref sig), who) => match ed25519::Public::try_from(&who.0[2..]) {
+			(Self::Ed25519(ref sig), who) => match ed25519::Public::try_from(who.as_ref()) {
 				Ok(signer) => sig.verify(msg, &signer),
 				Err(()) => false,
 			},
-			(Self::Sr25519(ref sig), who) => match sr25519::Public::try_from(&who.0[2..]) {
+			(Self::Sr25519(ref sig), who) => match sr25519::Public::try_from(who.as_ref()) {
 				Ok(signer) => sig.verify(msg, &signer),
 				Err(()) => false,
 			},
-			(Self::Secp256k1(ref sig), who) => match ecdsa::Public::try_from(&who.0[2..]) {
-				Ok(signer) => {
-					let m = sp_io::hashing::blake2_256(msg.get());
-					match sp_io::crypto::secp256k1_ecdsa_recover_compressed(sig.as_ref(), &m) {
-						Ok(pubkey) => pubkey == signer.0,
-						_ => false,
-					}
-				},
-				Err(_) => false,
+			(Self::Secp256k1(ref sig), who) => {
+				let m = sp_io::hashing::blake2_256(msg.get());
+				match sp_io::crypto::secp256k1_ecdsa_recover_compressed(sig.as_ref(), &m) {
+					Ok(pubkey) =>
+						&sp_io::hashing::blake2_256(pubkey.as_ref()) ==
+							<dyn AsRef<[u8; 32]>>::as_ref(who),
+					_ => false,
+				}
 			},
-			(Self::P256(ref sig), who) => match p256::Public::try_from(&who.0[2..]) {
-				Ok(signer) => np_io::crypto::p256_verify(sig, msg.get(), &signer),
-				Err(_) => false,
+			(Self::P256(ref sig), who) => {
+				let m = sp_io::hashing::blake2_256(msg.get());
+				match np_io::crypto::p256_recover_compressed(sig.as_ref(), &m) {
+					Some(pubkey) =>
+						&sp_io::hashing::blake2_256(pubkey.as_ref()) ==
+							<dyn AsRef<[u8; 32]>>::as_ref(who),
+					_ => false,
+				}
 			},
+			_ => false,
+			/*
+
 			(Self::WebAuthn(ref sig), who) => match p256::Public::try_from(&who.0[2..]) {
 				Ok(signer) => np_io::crypto::webauthn_verify(sig, msg.get(), &signer),
 				Err(_) => false,
 			},
+			*/
 		}
-	}
-}
-
-/// Public key for any known crypto algorithm.
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
-#[derive(Eq, PartialEq, Ord, PartialOrd, Clone, Encode, Decode, RuntimeDebug, TypeInfo)]
-pub enum UniversalSigner {
-	/// A Ed25519 identity.
-	Ed25519(ed25519::Public),
-	/// A Sr25519 identity.
-	Sr25519(sr25519::Public),
-	/// A Secp256k1 identity.
-	Secp256k1(ecdsa::Public),
-	/// A P-256 identity.
-	P256(p256::Public),
-}
-
-impl IdentifyAccount for UniversalSigner {
-	type AccountId = UniversalAddress;
-
-	fn into_account(self) -> Self::AccountId {
-		match self {
-			Self::Ed25519(k) => k.into(),
-			Self::Sr25519(k) => k.into(),
-			Self::Secp256k1(k) => k.into(),
-			Self::P256(k) => k.into(),
-		}
-	}
-}
-
-impl From<ed25519::Public> for UniversalSigner {
-	fn from(k: ed25519::Public) -> Self {
-		Self::Ed25519(k)
-	}
-}
-
-impl From<sr25519::Public> for UniversalSigner {
-	fn from(k: sr25519::Public) -> Self {
-		Self::Sr25519(k)
-	}
-}
-
-impl From<ecdsa::Public> for UniversalSigner {
-	fn from(k: ecdsa::Public) -> Self {
-		Self::Secp256k1(k)
-	}
-}
-
-impl From<p256::Public> for UniversalSigner {
-	fn from(k: p256::Public) -> Self {
-		Self::P256(k)
 	}
 }
